Add reusable LoRaWAN helpers
